added new contact us page on navbar
