+++ src/components/faq/images/discord-logo.png
